+++ src/main/kotlin/org/github/olloginov/ideataskbrowser/util/TaskHelper.kt
@@ -1,6 +1,9 @@
 package org.github.olloginov.ideataskbrowser.util
 
 import com.intellij.tasks.Task
+import com.intellij.tasks.gitlab.GitlabTask
+import com.intellij.tasks.gitlab.model.GitlabIssue
+import java.lang.reflect.InaccessibleObjectException
 import java.util.Date
 
 object TaskHelper {
@@ -13,5 +16,28 @@ object TaskHelper {
             return a ?: b
         }
         return if (a.after(b)) a else b
+    }
+
+    fun getPresentationTitle(task: Task): String {
+        return if (task.isIssue) {
+            task.presentableId + ": " + task.summary
+        } else {
+            task.summary
+        }
+    }
+
+    fun getGitlabDescription(task: GitlabTask): String? {
+        return try {
+            val field = task.javaClass.getDeclaredField("myIssue")
+            field.isAccessible = true
+            val issue = field.get(task) as GitlabIssue
+            issue.description
+        } catch (e: NoSuchFieldException) {
+            null
+        } catch (e: InaccessibleObjectException) {
+            null
+        } catch (e: SecurityException) {
+            null
+        }
     }
 }